Move files that are attached to the contraction methods
