+++ src/contraction/CMakeLists.txt
@@ -2,4 +2,6 @@ ADD_LIBRARY(contraction OBJECT
     contractGraph.c
     contractGraph_driver.cpp
     contract.cpp
+    ch_vertex.cpp
+    ch_edge.cpp
     )
